feat: add authentication routes and enhance user registration and login functionality

+++ src/route/parking_slots/index.ts
@@ -14,7 +14,8 @@ export const parking_slots_route = new Elysia({
 })
   .get(
     '/',
-    async ({ query }) => {
+    async ({ query, set }) => {
+      console.log('query', query);
       try {
         const { search, floor, status } = query;
         console.log(query);
@@ -47,6 +48,8 @@ export const parking_slots_route = new Elysia({
           },
         });
 
+        set.status = 200;
+
         return { data: parking_slots, status: 200 };
       } catch (e: any) {
         return { message: 'Internal Server Error' };
@@ -63,13 +66,14 @@ export const parking_slots_route = new Elysia({
 
   .post(
     '/',
-    async ({ body }) => {
+    async ({ body, set }) => {
       try {
-        const { floor_id, slot_number, status } = body;
+        const { floor_number, slot_number } = body;
 
         const validated_parking_slots = validate_create_parking_slot.safeParse(body);
 
         if (!validated_parking_slots.success) {
+          set.status = 400;
           return { message: validated_parking_slots.error.issues[0].message };
         }
 
@@ -80,43 +84,83 @@ export const parking_slots_route = new Elysia({
         });
 
         if (is_slot_number) {
+          set.status = 400;
           return {
             message: 'slot number already exits',
             status: 400,
           };
         }
 
-        const new_parking_slot = await prisma.parking_slots.create({
-          data: {
-            slot_number: slot_number,
-            floor_id: floor_id,
-            status: status as ParkingStatus,
+        const is_parking_in_floor = await prisma.parking_slots.findMany({
+          include: {
+            floor: true,
+          },
+          where: {
+            floor: {
+              floor_number: floor_number,
+            },
           },
         });
 
+        if (is_parking_in_floor.length >= 6) {
+          set.status = 400;
+          return {
+            message: 'parking slot in floor is full',
+            status: 400,
+          };
+        }
+
+        const floor = await prisma.floor.findFirst({
+          where: {
+            floor_number: floor_number,
+          },
+        });
+
+        if (!floor) {
+          set.status = 400;
+          return {
+            message: 'floor not found',
+            status: 400,
+          };
+        }
+        console.log('floor', floor);
+
+        const new_parking_slot = await prisma.parking_slots.create({
+          data: {
+            slot_number: slot_number,
+            floor_id: floor.id,
+          },
+        });
+
+        set.status = 200;
         return {
           message: 'parking slot created successfully',
           data: new_parking_slot,
           status: 200,
         };
       } catch (e: any) {
+        console.log(e);
         return { message: 'Internal Server Error' };
       }
     },
     {
       body: t.Object({
         slot_number: t.String(),
-        floor_id: t.String(),
-        status: t.String(),
+        floor_number: t.String(),
       }),
     }
   )
   .put(
-    '/',
-    async ({ body }) => {
-      try {
-        const { floor_id, slot_number, status, parking_slot_id } = body;
+    '/id/:parking_slot_id',
+    async ({ body, set, params }) => {
+      console.log('api');
 
+      try {
+        const { floor_number, slot_number } = body;
+        const { parking_slot_id } = params;
+        // console.log('floor_number', floor_number);
+        // console.log('slot_number', slot_number);
+        // console.log('parking_slot_id', parking_slot_id);
         const validated_parking_slots = validate_update_parking_slot.safeParse(body);
 
         if (!validated_parking_slots.success) {
@@ -130,8 +174,23 @@ export const parking_slots_route = new Elysia({
         });
 
         if (!is_slot_number) {
+          set.status = 400;
           return {
             message: 'parking slot not found',
+            status: 400,
+          };
+        }
+
+        const floor = await prisma.floor.findFirst({
+          where: {
+            floor_number: floor_number,
+          },
+        });
+
+        if (!floor) {
+          set.status = 400;
+          return {
+            message: 'floor not found',
             status: 400,
           };
         }
@@ -140,52 +199,63 @@ export const parking_slots_route = new Elysia({
           where: { id: parking_slot_id },
           data: {
             slot_number: slot_number,
-            floor_id: floor_id,
-            status: status as ParkingStatus,
+            floor_id: floor.id,
           },
         });
-
+        set.status = 200;
         return {
           message: 'parking slot created successfully',
           data: updated_parking_slot,
           status: 200,
         };
       } catch (e: any) {
+        console.log(e);
         return { message: 'Internal Server Error' };
       }
     },
     {
       body: t.Object({
         slot_number: t.String(),
+        floor_number: t.String(),
+      }),
+      params: t.Object({
         parking_slot_id: t.String(),
-        floor_id: t.String(),
-        status: t.String(),
       }),
     }
   )
   .delete(
     '/id/:parking_slot_id',
-    async ({ params }) => {
+    async ({ params, set }) => {
       try {
-        const { parking_slot_id } = await params;
+        const { parking_slot_id } = params;
+
         const is_parking_slot = await prisma.parking_slots.findUnique({
-          where: {
-            id: parking_slot_id,
-          },
+          where: { id: parking_slot_id },
         });
 
         if (!is_parking_slot) {
-          return { message: 'parking slot not found', status: 404 };
+          set.status = 404;
+          return { message: 'Parking slot not found' };
+        }
+
+        const linked_reservations = await prisma.reservations.findMany({
+          where: { parking_slot_id },
+        });
+
+        if (linked_reservations.length > 0) {
+          set.status = 400;
+          return { message: 'Cannot delete parking slot because it has linked reservations' };
         }
 
         await prisma.parking_slots.delete({
-          where: {
-            id: parking_slot_id,
-          },
+          where: { id: parking_slot_id },
         });
 
-        return { message: 'Delete parking slot success', status: 200 };
+        set.status = 200;
+        return { message: 'Delete parking slot success' };
       } catch (e: any) {
+        // console.log(e);
+        set.status = 500;
         return { message: 'Internal Server Error' };
       }
     },
